+++ src/models/operations/createoutput.ts
@@ -11176,17 +11176,6 @@ export type OutputTcpjson = {
   textSecret?: string | undefined;
 };
 
-export type UrlWizHec = {
-  /**
-   * URL to an endpoint to send events to, such as http://localhost:8088/services/collector/event
-   */
-  url: string;
-  /**
-   * Assign a weight (>0) to each endpoint to indicate its traffic-handling capability
-   */
-  weight?: number | undefined;
-};
-
 export type OutputWizHec = {
   /**
    * Unique ID for this output
@@ -11209,10 +11198,7 @@ export type OutputWizHec = {
    * Tags for filtering and grouping in @{product}
    */
   streamtags?: Array<string> | undefined;
-  /**
-   * Enable for optimal performance. Even if you have one hostname, it can expand to multiple IPs. If disabled, consider enabling round-robin DNS.
-   */
-  loadBalanced?: boolean | undefined;
+  loadBalanced?: any | undefined;
   /**
    * In the Splunk app, define which Splunk processing queue to send the events after HEC processing.
    */
@@ -11266,10 +11252,7 @@ export type OutputWizHec = {
    * List of headers that are safe to log in plain text
    */
   safeHeaders?: Array<string> | undefined;
-  /**
-   * Output metrics in multiple-metric format
-   */
-  enableMultiMetrics?: boolean | undefined;
+  enableMultiMetrics?: any | undefined;
   /**
    * Select Manual to enter an auth token directly, or select Secret to use a text secret to authenticate
    */
@@ -11289,28 +11272,20 @@ export type OutputWizHec = {
    * How to handle events when all receivers are exerting backpressure
    */
   onBackpressure?: models.BackpressureBehaviorOptions | undefined;
+  /**
+   * The unique identifier for the specific Cribl connector defined in your Wiz Settings. This is used to cross-validate the bearer token and ensure traffic is originating from the authorized integration.
+   */
+  wiz_connector_id: string;
+  /**
+   * Your Wiz deployment environment.
+   */
+  wiz_environment: string;
+  /**
+   * Your Wiz deployment data center (e.g., us1, us8, eu1). From Tenant Info → Data Center and Regions → Tenant Data Center in your Wiz console.
+   */
+  data_center: string;
+  wiz_sourcetype: string;
   description?: string | undefined;
-  /**
-   * URL to an endpoint to send events to, such as http://localhost:8088/services/collector/event
-   */
-  url?: string | undefined;
-  /**
-   * Enable round-robin DNS lookup. When a DNS server returns multiple addresses, @{product} will cycle through them in the order returned. For optimal performance, consider enabling this setting for non-load balanced destinations.
-   */
-  useRoundRobinDns?: boolean | undefined;
-  /**
-   * Exclude all IPs of the current host from the list of any resolved hostnames
-   */
-  excludeSelf?: boolean | undefined;
-  urls?: Array<UrlWizHec> | undefined;
-  /**
-   * The interval in which to re-resolve any hostnames and pick up destinations from A records
-   */
-  dnsResolvePeriodSec?: number | undefined;
-  /**
-   * How far back in time to keep traffic stats for load balancing purposes
-   */
-  loadBalanceStatsPeriodSec?: number | undefined;
   /**
    * Wiz Defender Auth token
    */
@@ -21158,26 +21133,6 @@ export function outputTcpjsonToJSON(outputTcpjson: OutputTcpjson): string {
 }
 
 /** @internal */
-export type UrlWizHec$Outbound = {
-  url: string;
-  weight?: number | undefined;
-};
-
-/** @internal */
-export const UrlWizHec$outboundSchema: z.ZodType<
-  UrlWizHec$Outbound,
-  z.ZodTypeDef,
-  UrlWizHec
-> = z.object({
-  url: z.string(),
-  weight: z.number().optional(),
-});
-
-export function urlWizHecToJSON(urlWizHec: UrlWizHec): string {
-  return JSON.stringify(UrlWizHec$outboundSchema.parse(urlWizHec));
-}
-
-/** @internal */
 export type OutputWizHec$Outbound = {
   id: string;
   type: "wiz_hec";
@@ -21185,7 +21140,7 @@ export type OutputWizHec$Outbound = {
   systemFields?: Array<string> | undefined;
   environment?: string | undefined;
   streamtags?: Array<string> | undefined;
-  loadBalanced?: boolean | undefined;
+  loadBalanced?: any | undefined;
   nextQueue?: string | undefined;
   tcpRouting?: string | undefined;
   tls?: models.TlsSettingsClientSideType1$Outbound | undefined;
@@ -21201,7 +21156,7 @@ export type OutputWizHec$Outbound = {
     | undefined;
   failedRequestLoggingMode?: string | undefined;
   safeHeaders?: Array<string> | undefined;
-  enableMultiMetrics?: boolean | undefined;
+  enableMultiMetrics?: any | undefined;
   authType?: string | undefined;
   responseRetrySettings?:
     | Array<models.ItemsTypeResponseRetrySettings$Outbound>
@@ -21209,13 +21164,11 @@ export type OutputWizHec$Outbound = {
   timeoutRetrySettings?: models.TimeoutRetrySettingsType$Outbound | undefined;
   responseHonorRetryAfterHeader?: boolean | undefined;
   onBackpressure?: string | undefined;
+  wiz_connector_id: string;
+  wiz_environment: string;
+  data_center: string;
+  wiz_sourcetype: string;
   description?: string | undefined;
-  url?: string | undefined;
-  useRoundRobinDns?: boolean | undefined;
-  excludeSelf?: boolean | undefined;
-  urls?: Array<UrlWizHec$Outbound> | undefined;
-  dnsResolvePeriodSec?: number | undefined;
-  loadBalanceStatsPeriodSec?: number | undefined;
   token?: string | undefined;
   textSecret?: string | undefined;
 };
@@ -21232,7 +21185,7 @@ export const OutputWizHec$outboundSchema: z.ZodType<
   systemFields: z.array(z.string()).optional(),
   environment: z.string().optional(),
   streamtags: z.array(z.string()).optional(),
-  loadBalanced: z.boolean().optional(),
+  loadBalanced: z.any().optional(),
   nextQueue: z.string().optional(),
   tcpRouting: z.string().optional(),
   tls: models.TlsSettingsClientSideType1$outboundSchema.optional(),
@@ -21248,7 +21201,7 @@ export const OutputWizHec$outboundSchema: z.ZodType<
   failedRequestLoggingMode: models
     .FailedRequestLoggingModeOptions$outboundSchema.optional(),
   safeHeaders: z.array(z.string()).optional(),
-  enableMultiMetrics: z.boolean().optional(),
+  enableMultiMetrics: z.any().optional(),
   authType: models.AuthenticationMethodOptionsAuthTokensItems$outboundSchema
     .optional(),
   responseRetrySettings: z.array(
@@ -21258,13 +21211,11 @@ export const OutputWizHec$outboundSchema: z.ZodType<
     .optional(),
   responseHonorRetryAfterHeader: z.boolean().optional(),
   onBackpressure: models.BackpressureBehaviorOptions$outboundSchema.optional(),
+  wiz_connector_id: z.string(),
+  wiz_environment: z.string(),
+  data_center: z.string(),
+  wiz_sourcetype: z.string(),
   description: z.string().optional(),
-  url: z.string().optional(),
-  useRoundRobinDns: z.boolean().optional(),
-  excludeSelf: z.boolean().optional(),
-  urls: z.array(z.lazy(() => UrlWizHec$outboundSchema)).optional(),
-  dnsResolvePeriodSec: z.number().optional(),
-  loadBalanceStatsPeriodSec: z.number().optional(),
   token: z.string().optional(),
   textSecret: z.string().optional(),
 });